dse: Reorganise modules by backend
