+++ devito/dse/backends/__init__.py
@@ -1,0 +1,6 @@
+from devito.dse.backends.common import *  # noqa
+from devito.dse.backends.basic import BasicRewriter  # noqa
+from devito.dse.backends.advanced import AdvancedRewriter  # noqa
+from devito.dse.backends.speculative import (SpeculativeRewriter,  # noqa
+                                             AggressiveRewriter,
+                                             CustomRewriter)
